src/main/webapp/app/entities/personnel-appui/update/personnel-appui-update.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { FormBuilder } from '@angular/forms';
import { ActivatedRoute } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of, Subject, from } from 'rxjs';

import { PersonnelAppuiService } from '../service/personnel-appui.service';
import { IPersonnelAppui, PersonnelAppui } from '../personnel-appui.model';
import { ILyceesTechniques } from 'app/entities/lycees-techniques/lycees-techniques.model';
import { LyceesTechniquesService } from 'app/entities/lycees-techniques/service/lycees-techniques.service';
import { IProviseur } from 'app/entities/proviseur/proviseur.model';
import { ProviseurService } from 'app/entities/proviseur/service/proviseur.service';

import { PersonnelAppuiUpdateComponent } from './personnel-appui-update.component';

describe('PersonnelAppui Management Update Component', () => {
  let comp: PersonnelAppuiUpdateComponent;
  let fixture: ComponentFixture<PersonnelAppuiUpdateComponent>;
  let activatedRoute: ActivatedRoute;
  let personnelAppuiService: PersonnelAppuiService;
  let lyceesTechniquesService: LyceesTechniquesService;
  let proviseurService: ProviseurService;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      declarations: [PersonnelAppuiUpdateComponent],
      providers: [
        FormBuilder,
        {
          provide: ActivatedRoute,
          useValue: {
            params: from([{}]),
          },
        },
      ],
    })
      .overrideTemplate(PersonnelAppuiUpdateComponent, '')
      .compileComponents();

    fixture = TestBed.createComponent(PersonnelAppuiUpdateComponent);
    activatedRoute = TestBed.inject(ActivatedRoute);
    personnelAppuiService = TestBed.inject(PersonnelAppuiService);
    lyceesTechniquesService = TestBed.inject(LyceesTechniquesService);
    proviseurService = TestBed.inject(ProviseurService);

    comp = fixture.componentInstance;
  });

  describe('ngOnInit', () => {
    it('Should call LyceesTechniques query and add missing value', () => {
      const personnelAppui: IPersonnelAppui = { id: 456 };
      const lyceesTechniques: ILyceesTechniques = { id: 33173 };
      personnelAppui.lyceesTechniques = lyceesTechniques;

      const lyceesTechniquesCollection: ILyceesTechniques[] = [{ id: 49271 }];
      jest.spyOn(lyceesTechniquesService, 'query').mockReturnValue(of(new HttpResponse({ body: lyceesTechniquesCollection })));
      const additionalLyceesTechniques = [lyceesTechniques];
      const expectedCollection: ILyceesTechniques[] = [...additionalLyceesTechniques, ...lyceesTechniquesCollection];
      jest.spyOn(lyceesTechniquesService, 'addLyceesTechniquesToCollectionIfMissing').mockReturnValue(expectedCollection);

      activatedRoute.data = of({ personnelAppui });
      comp.ngOnInit();

      expect(lyceesTechniquesService.query).toHaveBeenCalled();
      expect(lyceesTechniquesService.addLyceesTechniquesToCollectionIfMissing).toHaveBeenCalledWith(
        lyceesTechniquesCollection,
        ...additionalLyceesTechniques
      );
      expect(comp.lyceesTechniquesSharedCollection).toEqual(expectedCollection);
    });

    it('Should call Proviseur query and add missing value', () => {
      const personnelAppui: IPersonnelAppui = { id: 456 };
      const directeur: IProviseur = { id: 37874 };
      personnelAppui.directeur = directeur;

      const proviseurCollection: IProviseur[] = [{ id: 54611 }];
      jest.spyOn(proviseurService, 'query').mockReturnValue(of(new HttpResponse({ body: proviseurCollection })));
      const additionalProviseurs = [directeur];
      const expectedCollection: IProviseur[] = [...additionalProviseurs, ...proviseurCollection];
      jest.spyOn(proviseurService, 'addProviseurToCollectionIfMissing').mockReturnValue(expectedCollection);

      activatedRoute.data = of({ personnelAppui });
      comp.ngOnInit();

      expect(proviseurService.query).toHaveBeenCalled();
      expect(proviseurService.addProviseurToCollectionIfMissing).toHaveBeenCalledWith(proviseurCollection, ...additionalProviseurs);
      expect(comp.proviseursSharedCollection).toEqual(expectedCollection);
    });

    it('Should update editForm', () => {
      const personnelAppui: IPersonnelAppui = { id: 456 };
      const lyceesTechniques: ILyceesTechniques = { id: 4352 };
      personnelAppui.lyceesTechniques = lyceesTechniques;
      const directeur: IProviseur = { id: 57033 };
      personnelAppui.directeur = directeur;

      activatedRoute.data = of({ personnelAppui });
      comp.ngOnInit();

      expect(comp.editForm.value).toEqual(expect.objectContaining(personnelAppui));
      expect(comp.lyceesTechniquesSharedCollection).toContain(lyceesTechniques);
      expect(comp.proviseursSharedCollection).toContain(directeur);
    });
  });

  describe('save', () => {
    it('Should call update service on save for existing entity', () => {
      // GIVEN
      const saveSubject = new Subject<HttpResponse<PersonnelAppui>>();
      const personnelAppui = { id: 123 };
      jest.spyOn(personnelAppuiService, 'update').mockReturnValue(saveSubject);
      jest.spyOn(comp, 'previousState');
      activatedRoute.data = of({ personnelAppui });
      comp.ngOnInit();

      // WHEN
      comp.save();
      expect(comp.isSaving).toEqual(true);
      saveSubject.next(new HttpResponse({ body: personnelAppui }));
      saveSubject.complete();

      // THEN
      expect(comp.previousState).toHaveBeenCalled();
      expect(personnelAppuiService.update).toHaveBeenCalledWith(personnelAppui);
      expect(comp.isSaving).toEqual(false);
    });

    it('Should call create service on save for new entity', () => {
      // GIVEN
      const saveSubject = new Subject<HttpResponse<PersonnelAppui>>();
      const personnelAppui = new PersonnelAppui();
      jest.spyOn(personnelAppuiService, 'create').mockReturnValue(saveSubject);
      jest.spyOn(comp, 'previousState');
      activatedRoute.data = of({ personnelAppui });
      comp.ngOnInit();

      // WHEN
      comp.save();
      expect(comp.isSaving).toEqual(true);
      saveSubject.next(new HttpResponse({ body: personnelAppui }));
      saveSubject.complete();

      // THEN
      expect(personnelAppuiService.create).toHaveBeenCalledWith(personnelAppui);
      expect(comp.isSaving).toEqual(false);
      expect(comp.previousState).toHaveBeenCalled();
    });

    it('Should set isSaving to false on error', () => {
      // GIVEN
      const saveSubject = new Subject<HttpResponse<PersonnelAppui>>();
      const personnelAppui = { id: 123 };
      jest.spyOn(personnelAppuiService, 'update').mockReturnValue(saveSubject);
      jest.spyOn(comp, 'previousState');
      activatedRoute.data = of({ personnelAppui });
      comp.ngOnInit();

      // WHEN
      comp.save();
      expect(comp.isSaving).toEqual(true);
      saveSubject.error('This is an error!');

      // THEN
      expect(personnelAppuiService.update).toHaveBeenCalledWith(personnelAppui);
      expect(comp.isSaving).toEqual(false);
      expect(comp.previousState).not.toHaveBeenCalled();
    });
  });

  describe('Tracking relationships identifiers', () => {
    describe('trackLyceesTechniquesById', () => {
      it('Should return tracked LyceesTechniques primary key', () => {
        const entity = { id: 123 };
        const trackResult = comp.trackLyceesTechniquesById(0, entity);
        expect(trackResult).toEqual(entity.id);
      });
    });

    describe('trackProviseurById', () => {
      it('Should return tracked Proviseur primary key', () => {
        const entity = { id: 123 };
        const trackResult = comp.trackProviseurById(0, entity);
        expect(trackResult).toEqual(entity.id);
      });
    });
  });
});
